feat: support listing of blobs

+++ _test_util.ts
@@ -6,6 +6,7 @@ export { assertEquals } from "jsr:@std/assert@~1/equals";
 export { assertNotEquals } from "jsr:@std/assert@~1/not-equals";
 export { assertRejects } from "jsr:@std/assert@~1/rejects";
 export { assertStrictEquals } from "jsr:@std/assert@~1/strict-equals";
+export { assertThrows } from "jsr:@std/assert@~1/throws";
 export { timingSafeEqual } from "jsr:@std/crypto@~1/timing-safe-equal";
 
 let kv: { close(): void } | undefined;
@@ -16,12 +17,13 @@ export async function getPath() {
 }
 
 export async function setup() {
-  return kv = await Deno.openKv(await getPath());
+  return kv = await Deno.openKv(":memory:");
 }
 
-export function cleanup() {
-  assert(path);
-  return Deno.remove(path);
+export function cleanup(): Promise<void> | void {
+  if (path && path !== ":memory:") {
+    return Deno.remove(path);
+  }
 }
 
 export function teardown() {
